Change ModelResult interface .getFieldAtStep() to .getFieldAtOutputIndex().
It's much harder to generalise the comparison of 'step' from different simulators etc.  So here we decide to use index of output instead because we mostly compare result at either the initial (index 0) or the final (index -1) and it's hard for step to be comparable between simulators.  It is users responsibility now to ensure the result set are matching.